Smarter and more realistic login
